move draw item stack into RenderUtil
