+++ src/main/java/gregtech/client/utils/RenderUtil.java
@@ -720,6 +720,28 @@ public class RenderUtil {
         return getTextureMap().getMissingSprite();
     }
 
+    @SideOnly(Side.CLIENT)
+    public static void drawItemStack(ItemStack itemStack, int x, int y, @Nullable String altTxt) {
+        GlStateManager.pushMatrix();
+        GlStateManager.translate(0.0F, 0.0F, 32.0F);
+        GlStateManager.color(1F, 1F, 1F, 1F);
+        GlStateManager.enableDepth();
+        GlStateManager.enableRescaleNormal();
+        GlStateManager.enableLighting();
+        RenderHelper.enableGUIStandardItemLighting();
+        OpenGlHelper.setLightmapTextureCoords(OpenGlHelper.lightmapTexUnit, 240.0f, 240.0f);
+        Minecraft mc = Minecraft.getMinecraft();
+        RenderItem itemRender = mc.getRenderItem();
+        itemRender.renderItemAndEffectIntoGUI(itemStack, x, y);
+        itemRender.renderItemOverlayIntoGUI(mc.fontRenderer, itemStack, x, y, altTxt);
+        GlStateManager.disableRescaleNormal();
+        GlStateManager.disableLighting();
+        GlStateManager.color(1F, 1F, 1F, 1F);
+        GlStateManager.popMatrix();
+        GlStateManager.enableBlend();
+        GlStateManager.disableDepth();
+    }
+
     public static void drawSlotOverlay(@NotNull IWidget slot, int overlayColor) {
         GlStateManager.colorMask(true, true, true, false);
         GuiDraw.drawRect(1, 1, slot.getArea().w() - 2, slot.getArea().h() - 2, overlayColor);
